make: got rid of strict aliasing warnings
Unfortunately I still don't know a good/performant way to get rid of the
strict aliasing warnings on newer GCC's. Creating a union is not really
desirable as it's been reported to lower performance and creates a lot more
typing work. The cases that are in the codebase now are false positives
though.

+++ src/main.c
@@ -578,7 +578,7 @@ int main(int argc, char* argv[]) {
         GLuint64 timerRend;
         glGetQueryObjectui64v(gQueries[gQueryFrontBuffer][TIMER_RENDER], GL_QUERY_RESULT, &timerRend);
 
-        printf("render objects: %f ms, swap buffers: %f\n", timerRend / 1000000.0, (double) elapsedSwap);
+        printf("render objects: %f ms, swap buffers: %f\n", (double) timerRend / 1000000.0, (double) elapsedSwap);
 
         /* swap the query set */
         swapQueryBuffers();
